feat: Use threads instead of forks
This changes the underlying concurrency model of our program. Instead of
forking the program into one half that sends messages and another half
that retrieves them, we do it all in one process.

This change was made to make it easier to keep track of which resources
the program was taking, and to simplify the process of freeing those
resources.