fix: Reverse order of threads
Previously, `receive_messages` was in a pthread, and `handle_events` was
the main thread. The problem with this was that `receive_messages` is
the thread that gets the exit signal from the server, and had to kill
the "parent" process in order to exit the process, which had bad
implications for handling freeing resources.

This way is just better.